refactor mapbox client
